refactor(beam): extract a shared Generator base for comprehension qualifiers
The `generate`, `b_generate` and `m_generate` qualifier objects each duplicated the same `{Tag, Anno, Pattern, Expression}` decoding and `pattern <- expression` scope threading in `toMacroStringDeclaredScope`.

Extract an abstract `Generator(tag, expressionDescription)` base that owns the shared decoding and scope handling. Subtypes now supply only what differs: the pattern rendering and, where the surrounding syntax is not a plain `pattern <- expression` (bitstring), how pattern and expression are combined.
